adjust themestyles, continue building StatusMachine
